Split effect contracts and gate builder append
